requirements.txt
aiodns==2.0.0
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.1.0
cffi==1.12.3
chardet==3.0.4
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
pycares==3.0.0
pycparser==2.19
slackclient==2.0.1
typing==3.7.4
typing-extensions==3.7.4
yarl==1.3.0
